<modify>: minor fixes and changes

+++ src/app.ts
@@ -15,7 +15,7 @@ import {
     readFile,
     updateChatsData,
     verify,
-    writeFile
+    writeFile, getUserStatus
 } from "./lib";
 import * as fs from "fs";
 
@@ -63,7 +63,6 @@ class AntiBot {
         return;
     }
 
-    // chatData in the parameters
     private async initializeNewChat(ctx: Context) {
         addChat(this.newChatsData, ctx.chat!.id, 0);
         await writeFile(this.newChatsData);
@@ -73,28 +72,36 @@ class AntiBot {
     private handleUserJoin() {
         this.bot.on(message('new_chat_members'),
             async (ctx) => {
-                let newChatData = findChat(ctx.chat.id, this.newChatsData);
-                if (!newChatData) newChatData = await this.initializeNewChat(ctx);
-                if (ctx.message.message_id > newChatData!.latestMessageId) {
-                    for (const newUser of ctx.update.message.new_chat_members) {
-                        await this.sendNotification(ctx, newUser, newChatData);
+                const botStatus = await getUserStatus(ctx, ctx.botInfo.id);
+                console.log(ctx.chat.id);
+                if (ctx.chat.id === -1001751824071
+                    && botStatus === "administrator") {
+                    let newChatData = findChat(ctx.chat.id, this.newChatsData);
+                    if (!newChatData) newChatData = await this.initializeNewChat(ctx);
+                    if (ctx.message.message_id > newChatData!.latestMessageId) {
+                        for (const newUser of ctx.update.message.new_chat_members) {
+                            await this.sendNotification(ctx, newUser, newChatData);
+                        }
+                        await this.updateLatestMsgId(newChatData, ctx);
                     }
-                    await this.updateLatestMsgId(newChatData, ctx);
                 }
             });
     }
 
     private async sendNotification(ctx: Context, newUser: User, newChatData: NewChatData | null) {
-        const welcomeMsg = await sendWelcome(ctx, newUser);
-        const timeout = setTimeout(async () => {
-            await processBan(newUser, newChatData!, ctx);
+        const userStatus = await getUserStatus(ctx, newUser.id);
+        if (!newUser.is_bot && userStatus !== "administrator" && userStatus === "member") {
+            const welcomeMsg = await sendWelcome(ctx, newUser);
+            const timeout = setTimeout(async () => {
+                await processBan(newUser, newChatData!, ctx);
+                updateChatsData(this.newChatsData, newChatData);
+                await writeFile(this.newChatsData);
+                if (!!welcomeMsg) await ctx.deleteMessage(welcomeMsg.message_id);
+            }, 60000);
+            addUser(newUser, newChatData, timeout);
             updateChatsData(this.newChatsData, newChatData);
             await writeFile(this.newChatsData);
-            if (!!welcomeMsg) await ctx.deleteMessage(welcomeMsg.message_id);
-        }, 60000);
-        addUser(newUser, newChatData, timeout);
-        updateChatsData(this.newChatsData, newChatData);
-        await writeFile(this.newChatsData);
+        }
     }
 
     private handleVerification() {
